Enhance expense creation with user association

+++ server/controllers/expense.controller.js
@@ -2,7 +2,12 @@ import Expense from '../models/expense.js'
 
 const createExpense = async (req, res) => {
     try {
-        const newExpense = new Expense(req.body)
+        const userId = req.user.userId
+        const expenseData = {
+            ...req.body,
+            user: userId,
+        }
+        const newExpense = new Expense(expenseData)
         const savedExpense = await newExpense.save()
         res.status(201).json(savedExpense)
     } catch (error) {
@@ -12,7 +17,7 @@ const createExpense = async (req, res) => {
 
 const getAllExpenses = async (req, res) => {
     try {
-        const expenses = await Expense.find()
+        const expenses = await Expense.find({ user: req.user.userId })
         res.status(200).json(expenses)
     } catch (error) {
         res.status(404).json({ message: error.message })
@@ -21,7 +26,7 @@ const getAllExpenses = async (req, res) => {
 
 const getExpenseById = async (req, res) => {
     try {
-        const expense = await Expense.findById(req.params.id)
+        const expense = await Expense.findById(req.params.id, { user: req.user.userId })
         if (!expense) return res.status(404).json({ message: "Expense not found" })
         res.status(200).json(expense)
     } catch (error) {
@@ -31,7 +36,7 @@ const getExpenseById = async (req, res) => {
 
 const updateExpenseById = async (req, res) => {
     try {
-        const updatedExpense = await Expense.findByIdAndUpdate(req.params.id, req.body, { new: true })
+        const updatedExpense = await Expense.findByIdAndUpdate(req.params.id, req.body, { new: true }, { user: req.user.userId })
         res.status(200).json(updatedExpense)
     } catch (error) {
         res.status(400).json({ message: error.message })
@@ -40,7 +45,7 @@ const updateExpenseById = async (req, res) => {
 
 const deleteExpenseById = async (req, res) => {
     try {
-        await Expense.findByIdAndDelete(req.params.id)
+        await Expense.findByIdAndDelete(req.params.id, { user: req.user.userId })
         res.status(204).send()
     } catch (error) {
         res.status(404).json({ message: error.message })
